# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dominion Realm API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/books": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "List Books",
        "operationId": "list_books_books_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookOut"
                  },
                  "type": "array",
                  "title": "Response List Books Books Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "books"
        ],
        "summary": "Create Book",
        "operationId": "create_book_books_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/books/{book_id}/manuscript": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "Manuscript",
        "operationId": "manuscript_books__book_id__manuscript_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManuscriptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/books/{book_id}/scenes/clear-draft": {
      "post": {
        "tags": [
          "books"
        ],
        "summary": "Clear Draft Scenes",
        "description": "Delete all non-approved scenes for a book (optional chapter scope). Approved prose is kept.",
        "operationId": "clear_draft_scenes_books__book_id__scenes_clear_draft_post",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Book Id"
            }
          },
          {
            "name": "chapter_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearDraftScenesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookIn": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "premise": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Premise"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "BookIn",
        "description": "POST body to create a book."
      },
      "BookOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "premise": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Premise"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "created_at"
        ],
        "title": "BookOut"
      },
...
      "ClearDraftScenesOut": {
        "properties": {
          "purged": {
            "type": "integer",
            "title": "Purged",
            "default": 0
          },
          "jobs_purged": {
            "type": "integer",
            "title": "Jobs Purged",
            "default": 0
          }
        },
        "type": "object",
        "title": "ClearDraftScenesOut",
        "description": "Result of removing all non-approved scenes (draft compile reset)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManuscriptChapter": {
        "properties": {
          "chapter_no": {
            "type": "integer",
            "title": "Chapter No"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "pov": {
            "type": "string",
            "title": "Pov"
          },
          "scenes": {
            "items": {
              "$ref": "#/components/schemas/ManuscriptScene"
            },
            "type": "array",
            "title": "Scenes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "chapter_no",
          "pov"
        ],
        "title": "ManuscriptChapter"
      },
      "ManuscriptOut": {
        "properties": {
          "book_id": {
            "type": "string",
            "format": "uuid",
            "title": "Book Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/ManuscriptChapter"
            },
            "type": "array",
            "title": "Chapters",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "book_id",
          "title"
        ],
        "title": "ManuscriptOut",
        "description": "The approved manuscript, assembled in reading order (latest approved version per scene)."
      },
      "ManuscriptScene": {
        "properties": {
          "scene_no": {
            "type": "integer",
            "title": "Scene No"
          },
          "prose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prose"
          }
        },
        "type": "object",
        "required": [
          "scene_no"
        ],
        "title": "ManuscriptScene"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}